Fix sink bug, code cleanup
